Restructured module.
