+++ pyVHDLModel/Interface.py
@@ -1,0 +1,128 @@
+from typing import Iterable
+
+from pyTooling.Decorators import export
+
+from pyVHDLModel import DocumentedEntityMixin, Mode, SubtypeOrSymbol, ExpressionUnion
+from pyVHDLModel.Object import Constant, Signal, Variable, File
+from pyVHDLModel.Subprogram import Procedure, Function
+from pyVHDLModel.Type import Type
+
+
+@export
+class InterfaceItem(DocumentedEntityMixin):
+	"""An ``InterfaceItem`` is a base-class for all mixin-classes for all interface items."""
+
+	def __init__(self, documentation: str = None):
+		super().__init__(documentation)
+
+
+@export
+class InterfaceItemWithMode:
+	"""An ``InterfaceItemWithMode`` is a mixin-class to provide a ``Mode`` to interface items."""
+
+	_mode: Mode
+
+	def __init__(self, mode: Mode):
+		self._mode = mode
+
+	@property
+	def Mode(self) -> Mode:
+		return self._mode
+
+
+@export
+class GenericInterfaceItem(InterfaceItem):
+	"""A ``GenericInterfaceItem`` is a mixin class for all generic interface items."""
+
+
+@export
+class PortInterfaceItem(InterfaceItem, InterfaceItemWithMode):
+	"""A ``PortInterfaceItem`` is a mixin class for all port interface items."""
+
+	def __init__(self, mode: Mode):
+		super().__init__()
+		InterfaceItemWithMode.__init__(self, mode)
+
+
+@export
+class ParameterInterfaceItem(InterfaceItem):
+	"""A ``ParameterInterfaceItem`` is a mixin class for all parameter interface items."""
+
+
+@export
+class GenericConstantInterfaceItem(Constant, GenericInterfaceItem, InterfaceItemWithMode):
+	def __init__(self, identifiers: Iterable[str], mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
+		super().__init__(identifiers, subtype, defaultExpression, documentation)
+		GenericInterfaceItem.__init__(self)
+		InterfaceItemWithMode.__init__(self, mode)
+
+
+@export
+class GenericTypeInterfaceItem(Type, GenericInterfaceItem):
+	def __init__(self, identifier: str, documentation: str = None):
+		super().__init__(identifier, documentation)
+		GenericInterfaceItem.__init__(self)
+
+
+@export
+class GenericSubprogramInterfaceItem(GenericInterfaceItem):
+	pass
+
+
+@export
+class GenericProcedureInterfaceItem(Procedure, GenericInterfaceItem):
+	def __init__(self, identifier: str, documentation: str = None):
+		super().__init__(identifier, documentation)
+		GenericInterfaceItem.__init__(self)
+
+
+@export
+class GenericFunctionInterfaceItem(Function, GenericInterfaceItem):
+	def __init__(self, identifier: str, documentation: str = None):
+		super().__init__(identifier, documentation)
+		GenericInterfaceItem.__init__(self)
+
+
+@export
+class GenericPackageInterfaceItem(GenericInterfaceItem):
+	def __init__(self, identifier: str, documentation: str = None):
+		#	super().__init__(identifier, documentation)
+		GenericInterfaceItem.__init__(self)
+
+
+@export
+class PortSignalInterfaceItem(Signal, PortInterfaceItem):
+	def __init__(self, identifiers: Iterable[str], mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
+		super().__init__(identifiers, subtype, defaultExpression, documentation)
+		PortInterfaceItem.__init__(self, mode)
+
+
+@export
+class ParameterConstantInterfaceItem(Constant, ParameterInterfaceItem, InterfaceItemWithMode):
+	def __init__(self, identifiers: Iterable[str], mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
+		super().__init__(identifiers, subtype, defaultExpression, documentation)
+		ParameterInterfaceItem.__init__(self)
+		InterfaceItemWithMode.__init__(self, mode)
+
+
+@export
+class ParameterVariableInterfaceItem(Variable, ParameterInterfaceItem, InterfaceItemWithMode):
+	def __init__(self, identifiers: Iterable[str], mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
+		super().__init__(identifiers, subtype, defaultExpression, documentation)
+		ParameterInterfaceItem.__init__(self)
+		InterfaceItemWithMode.__init__(self, mode)
+
+
+@export
+class ParameterSignalInterfaceItem(Signal, ParameterInterfaceItem, InterfaceItemWithMode):
+	def __init__(self, identifiers: Iterable[str], mode: Mode, subtype: SubtypeOrSymbol, defaultExpression: ExpressionUnion = None, documentation: str = None):
+		super().__init__(identifiers, subtype, defaultExpression, documentation)
+		ParameterInterfaceItem.__init__(self)
+		InterfaceItemWithMode.__init__(self, mode)
+
+
+@export
+class ParameterFileInterfaceItem(File, ParameterInterfaceItem):
+	def __init__(self, identifiers: Iterable[str], subtype: SubtypeOrSymbol, documentation: str = None):
+		super().__init__(identifiers, subtype, documentation)
+		ParameterInterfaceItem.__init__(self)
